refactor(compute): remove Passive channel and integrate IndependentIntegration into Markov
